+++ src/sklearn_evaluation/__init__.py
@@ -1,5 +1,7 @@
 __version__ = '0.6.1dev'
 
+from ploomber_core.telemetry.telemetry import Telemetry
+
 from .evaluator import ClassifierEvaluator
 from .nb.NotebookIntrospector import NotebookIntrospector
 from .nb.NotebookCollection import NotebookCollection
@@ -13,3 +15,11 @@ __all__ = [
     'NotebookCollection',
     'NotebookDatabase',
 ]
+
+telemetry = Telemetry(
+    api_key="phc_3Of87HTxtAH0GnqU5ApdJnUWUcDhaR11LuEHtVCV41Z",
+    package_name="sklearn-evaluation",
+    version=__version__,
+)
+
+telemetry.log_api('imported')