+++ src/sklearn_evaluation/plot/__init__.py
@@ -1,8 +1,8 @@
 """
 Plotting functions
 """
-from .classification import (confusion_matrix,
-                             feature_importances,
+
+from .classification import (confusion_matrix, feature_importances,
                              precision_at_proportions)
 
 from .roc import roc
@@ -12,7 +12,8 @@ from .validation_curve import validation_curve
 from .learning_curve import learning_curve
 from .metrics import metrics_at_thresholds
 
-
-__all__ = ['confusion_matrix', 'feature_importances', 'precision_recall',
-           'roc', 'precision_at_proportions', 'grid_search',
-           'validation_curve', 'learning_curve', 'metrics_at_thresholds']
+__all__ = [
+    'confusion_matrix', 'feature_importances', 'precision_recall', 'roc',
+    'precision_at_proportions', 'grid_search', 'validation_curve',
+    'learning_curve', 'metrics_at_thresholds'
+]